AKH
Adding some example code that works for adding a note to an order.
Needs a lot of clean u:
It seems to return an erropr even though it's successful
I've put a global variable in there to keep a hold of the access token which is pretty bad
There's some getting orders code that doesn't work
Need to make sure that the params you pass in actually get passed to the request and it's not hard coded
But I'm 1hr40mins past the time I should have left so I'm calling it.

PEACE